Factor out LumenOutput and SQLOutput
