+++ lumen/ai/prompts/check_validity.jinja2
@@ -1,7 +1,4 @@
-Based on the latest user's query, decide whether the
-current table and data contain all the required info to
-provide a reliable, accurate answer to the user's query.
-Return `is_invalid=True` if not; else return `is_invalid=False`.
+Based on the latest user's query, decide whether the current table and data contains the the required data. If the user is referencing "this data" or "that" they probably are referring to the current data. Pay particular attention to whether the data actually contains the columns required to answer the query, e.g. if they are asking for a location but there are no location related column the data is invalid.
 
 ### Current Table:
 ```
