+++ lumen/ai/utils.py
@@ -6,6 +6,7 @@ from functools import wraps
 from pathlib import Path
 
 import jinja2
+import pandas as pd
 
 from lumen.pipeline import Pipeline
 from lumen.sources.base import Source
@@ -104,3 +105,76 @@ def get_schema(
                 spec.pop("inclusiveMaximum")
     schema = format_schema(schema)
     return schema
+
+
+def describe_data(df: pd.DataFrame) -> str:
+    def format_float(num):
+        if pd.isna(num):
+            return num
+        # if is integer, round to 0 decimals
+        if num == int(num):
+            return f"{int(num)}"
+        elif 0.01 <= abs(num) < 100:
+            return f"{num:.1f}"  # Regular floating-point notation with two decimals
+        else:
+            return f"{num:.1e}"  # Exponential notation with two decimals
+
+    size = df.size
+    shape = df.shape
+    if size < 250:
+        return df
+
+    is_summarized = False
+    if shape[0] > 5000:
+        is_summarized = True
+        df = df.sample(5000)
+
+    df = df.sort_index()
+
+    for col in df.columns:
+        if isinstance(df[col].iloc[0], pd.Timestamp):
+            df[col] = pd.to_datetime(df[col])
+
+    df_describe_dict = df.describe(percentiles=[]).drop(["min", "max"]).to_dict()
+
+    for col in df.select_dtypes(include=["object"]).columns:
+        if col not in df_describe_dict:
+            df_describe_dict[col] = {}
+        df_describe_dict[col]["nunique"] = df[col].nunique()
+        try:
+            df_describe_dict[col]["lengths"] = {
+                "max": df[col].str.len().max(),
+                "min": df[col].str.len().min(),
+                "mean": float(df[col].str.len().mean()),
+            }
+        except AttributeError:
+            pass
+
+    for col in df.columns:
+        if col not in df_describe_dict:
+            df_describe_dict[col] = {}
+        df_describe_dict[col]["nulls"] = int(df[col].isnull().sum())
+
+    # select datetime64 columns
+    for col in df.select_dtypes(include=["datetime64"]).columns:
+        for key in df_describe_dict[col]:
+            df_describe_dict[col][key] = str(df_describe_dict[col][key])
+        df[col] = df[col].astype(str)  # shorten output
+
+    # select all numeric columns and round
+    for col in df.select_dtypes(include=["int64", "float64"]).columns:
+        for key in df_describe_dict[col]:
+            df_describe_dict[col][key] = format_float(df_describe_dict[col][key])
+
+    for col in df.select_dtypes(include=["float64"]).columns:
+        df[col] = df[col].apply(format_float)
+
+    data = {
+        "summary": {
+            "total_table_cells": size,
+            "total_shape": shape,
+            "is_summarized": is_summarized,
+        },
+        "stats": df_describe_dict,
+    }
+    return data
